requirements.txt
asgiref
asyncio
cachetools
certifi
charset-normalizer
colorama
decorator
Django
django-cors-headers
djangorestframework
google
google-api-core
google-auth
grpcio
grpcio-status
idna
proglog
proto-plus
protobuf
pyasn1
pyasn1-modules
pytz
requests
rsa
six
soupsieve
sqlparse
tqdm
tzdata
urllib3
gunicorn